fix: 重构点位的接口

+++ routes/point.js
@@ -30,29 +30,32 @@ const PointModel = require("../models/PointModel");
  * @apiSampleRequest http://localhost:5050/point/add
  * @apiVersion 1.0.0
  */
-router.post("/add", (req, res) => {
-  const {
-    port_point_map_id,
-    content
-  } = req.body;
-  PointModel.create({
-      content: content,
-      port_point_map_id: port_point_map_id,
+router.post("/add", async (req, res) => {
+  try {
+    const {
+      port_point_map_id,
+      content
+    } = req.body;
+    if (!port_point_map_id) {
+      return res.send(R.fail("港口点位图id不可以为空."))
+    }
+    if (!content) {
+      return res.send(R.fail("点位内容不可以为空."))
+    }
+    const {
+      dataValues
+    } = await PointModel.create({
+      content,
+      port_point_map_id,
     })
-    .then((point) => {
-      res.send({
-        status: 200,
-        msg: "添加点位成功.",
-        data: point,
-      });
-    })
-    .catch((error) => {
-      console.error("添加点位失败.", error);
-      res.send({
-        status: 400,
-        msg: "添加点位失败，请重试！",
-      });
-    });
+    if (dataValues !== null) {
+      return res.send(R.success(dataValues, "点位添加成功."))
+    } else {
+      return res.send(R.fail("点位添加失败."))
+    }
+  } catch (error) {
+    return res.send(R.fail("点位添加失败."))
+  }
 });
 
 /**
@@ -71,28 +74,23 @@ router.post("/add", (req, res) => {
  * @apiSampleRequest http://localhost:5050/point/delete
  * @apiVersion 1.0.0
  */
-router.get("/delete", (req, res) => {
+router.get("/delete", async (req, res) => {
   const {
     id
   } = req.query;
-  PointModel.destroy({
-      where: {
-        id: id,
-      },
-    })
-    .then((point) => {
-      res.send({
-        status: 200,
-        msg: "删除点位成功.",
-      });
-    })
-    .catch((error) => {
-      console.error("删除点位失败.", error);
-      res.send({
-        status: 400,
-        msg: "删除点位失败，请重试！",
-      });
-    });
+  if (!id) {
+    return res.send(R.fail("id不可以为空."))
+  }
+  const point = await PointModel.destroy({
+    where: {
+      id: id,
+    },
+  })
+  if (point) {
+    return res.send(R.success(point, "点位添加成功."))
+  } else {
+    return res.send(R.fail("删除点位失败."))
+  }
 });
 
 /**
@@ -114,26 +112,22 @@ router.get("/delete", (req, res) => {
  * @apiSampleRequest http://localhost:5050/point/update
  * @apiVersion 1.0.0
  */
-router.post("/update", (req, res) => {
-  const point = req.body;
-  PointModel.update(point, {
+router.post("/update", async (req, res) => {
+  try {
+    const point = req.body;
+    const [result] = await PointModel.update(point, {
       where: {
         id: point.id,
       },
     })
-    .then((point) => {
-      res.send({
-        status: 200,
-        msg: "修改点位成功.",
-      });
-    })
-    .catch((error) => {
-      console.error("修改点位失败.", error);
-      res.send({
-        status: 400,
-        msg: "修改点位失败，请重试！",
-      });
-    });
+    if (result) {
+      return res.send(R.success(point, "点位修改成功."))
+    } else {
+      return res.send(R.fail("点位修改失败."))
+    }
+  } catch (error) {
+    return res.send(R.fail("点位修改失败."))
+  }
 });
 
 /**
@@ -152,14 +146,15 @@ router.post("/update", (req, res) => {
  * @apiSampleRequest http://localhost:5050/point/search
  * @apiVersion 1.0.0
  */
-router.post("/search", (req, res) => {
-  const {
-    pageNum,
-    pageSize,
-    port_point_map_id,
-    content
-  } = req.body;
-  PointModel.findAll({
+router.post("/search", async (req, res) => {
+  try {
+    const {
+      pageNum,
+      pageSize,
+      port_point_map_id,
+      content
+    } = req.body;
+    const point = await PointModel.findAll({
       where: {
         [Op.or]: [{
             port_point_map_id: port_point_map_id ? port_point_map_id : "",
@@ -173,20 +168,14 @@ router.post("/search", (req, res) => {
         ["create_time"]
       ],
     })
-    .then((point) => {
-      res.send({
-        status: 200,
-        msg: "查询点位成功.",
-        data: utils.pageFilter(point, pageNum, pageSize),
-      });
-    })
-    .catch((error) => {
-      console.error("查询点位失败.", error);
-      res.send({
-        status: 400,
-        msg: "查询点位失败，请重试！",
-      });
-    });
+    if (point.length > 0) {
+      return res.send(R.success(utils.pageFilter(point, pageNum, pageSize), "查询点位成功."))
+    } else {
+      return res.send(R.fail("查询点位失败."))
+    }
+  } catch (error) {
+    return res.send(R.fail("查询点位失败."))
+  }
 });
 
 /**
@@ -206,30 +195,21 @@ router.post("/search", (req, res) => {
  * @apiSampleRequest http://localhost:5050/point/findAll
  * @apiVersion 1.0.0
  */
-router.post("/findAll", (req, res) => {
+router.post("/findAll", async (req, res) => {
   const {
     pageNum,
     pageSize
   } = req.body;
-  PointModel.findAll({
-      order: [
-        ["create_time", "DESC"]
-      ],
-    })
-    .then((point) => {
-      res.send({
-        status: 200,
-        msg: "获取所有点位图成功.",
-        data: utils.pageFilter(point, pageNum, pageSize),
-      });
-    })
-    .catch((err) => {
-      console.error("获取所有点位图失败.", err);
-      res.send({
-        status: 400,
-        msg: "获取所有点位图失败.",
-      });
-    });
+  const point = await PointModel.findAll({
+    order: [
+      ["create_time", "DESC"]
+    ],
+  })
+  if (point.length > 0) {
+    return res.send(R.success(utils.pageFilter(point, pageNum, pageSize), "获取所有点位成功."))
+  } else {
+    return res.send(R.fail("获取所有点位失败."))
+  }
 });
 
 /**
@@ -251,39 +231,21 @@ router.post("/batch/delete", async (req, res) => {
   const {
     pointIds
   } = req.body;
-  if (!pointIds) {
-    return res.send({
-      status: 400,
-      msg: "pointIds不可以为空",
-    });
+  if (pointIds.length <= 0) {
+    return res.send(R.fail("pointIds不可以为空."))
   }
-  await PointModel.destroy({
-      where: {
-        id: {
-          [Op.in]: pointIds,
-        },
+  const point = await PointModel.destroy({
+    where: {
+      id: {
+        [Op.in]: pointIds,
       },
-    })
-    .then((point) => {
-      if (point) {
-        res.send({
-          status: 200,
-          msg: "点位图批量删除成功.",
-        });
-      } else {
-        res.send({
-          status: 400,
-          msg: "点位图批量删除失败.",
-        });
-      }
-    })
-    .catch((err) => {
-      console.error("点位图批量删除失败.", err);
-      res.send({
-        status: 400,
-        msg: "点位图批量删除失败.",
-      });
-    });
+    },
+  })
+  if (point) {
+    return res.send(R.success(point, "批量删除点位成功."))
+  } else {
+    return res.send(R.fail("点位图批量删除失败."))
+  }
 });
 
 module.exports = router;
